pyproject.toml
[project]
name = "elyx"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.0",
    "certifi>=2025.8.3",
    "chromadb>=0.5.23",
    "crewai>=0.159.0",
    "crewai-tools>=0.62.0",
    "embedchain>=0.1.128",
    "faiss-cpu>=1.12.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-groq>=0.3.7",
    "ollama>=0.5.3",
    "plotly>=6.3.0",
    "python-dotenv>=1.1.1",
    "schedule>=1.2.2",
    "sentence-transformers>=5.1.0",
    "streamlit>=1.48.1",
    "tavily-python>=0.7.10",
    "urllib3>=2.5.0",
]
